Allow sizeof(void*) != sizeof(int).
--HG--
extra : convert_revision : svn%3A0fb12191-db4b-0410-8866-e78deaee71e5/trunk%40412

+++ mpqc/src/lib/Attic/comm/picl/shm/piclshmnode.c
@@ -487,7 +487,7 @@ recv0(buf,bytes,type)
    *    (int)((char*)lastmessage)-(int)((char*)NEXT_MESSAGE(message)));
    */
   memmove(message,NEXT_MESSAGE(message),
-        (int)((char*)lastmessage)-(int)((char*)NEXT_MESSAGE(message)));
+        (size_t)((char*)lastmessage)-(size_t)((char*)NEXT_MESSAGE(message)));
 
   commbuf[mynodeid]->nmsg--;
